requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
SQLAlchemy>=2.0.30,<3
pydantic>=2.7,<3
pydantic-settings>=2.2,<3
psycopg[binary]>=3.2,<4
PyJWT>=2.9,<3
alembic>=1.13,<2
pytest>=8.3,<9
httpx>=0.28,<1
